pyproject.toml
[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"


[project]
name = "oil-market-simulation"
version = "0.0.1"
description = "A Descrete-Time Agent-Based simulation of market shocks in the oil market."
authors = [
    {name = "Adrian Edwards", email = "aedwrds0603@gmail.com"}
]
requires-python = ">=3.10,<3.12"
readme = {file = "README.txt", content-type = "text/markdown"}

dependencies = [
    "pyyaml>=6.0",
    "numpy>=1.26",
    "pandas>=2.1",
    "matplotlib>=3.8",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]